chore: ESLint config, 61 lint fixes across 22 client files, CLAUDE.md update
- Add ESLint v9 flat config with typescript-eslint
- Fix no-floating-promises, no-unused-vars, consistent-return across client
- Refactor livekitSession: delegate entirely to AudioPipeline (1438→1171 lines)
- Add 7 delete-account UI tests in settings-overlay.test.ts
- Update CLAUDE.md with latest features and project structure
- Update .gitignore

+++ Client/tauri-client/src/components/message-list/fenwick.ts
@@ -20,7 +20,7 @@ export class FenwickTree {
     if (delta === 0) return;
     this.values[i] = value;
     for (let x = i + 1; x <= this.size; x += x & (-x)) {
-      (this.tree as Float64Array)[x] = (this.tree[x] as number) + delta;
+      (this.tree)[x] = (this.tree[x] as number) + delta;
     }
   }
